gcs: zero-extend HRESULT codes in rpcError.Unwrap (#2849)
Guest RPC failures carry their HRESULT in an int32 field. rpcError.Unwrap
converted this directly to a windows.Errno (uintptr), which sign-extends
any code with the high bit set (HRESULT failure codes, 0x8xxxxxxx and
above) into 0xFFFFFFFF_xxxxxxxx. That value no longer equals the canonical
syscall.Errno constants, so errors.Is comparisons against them - and every
helper built on top (e.g. hcs.IsNotExist and the "resource already gone"
checks used by teardown paths) - silently failed to match.

Convert the result through uint32 before widening so the code stays
zero-extended and matches the canonical constants for any HRESULT.

For example, with result = 0xc037010e (negative as an int32):

  before: windows.Errno(err.result)         => 0xFFFFFFFF_C037010E
  after:  windows.Errno(uint32(err.result)) => 0x00000000_C037010E

  errors.Is(err, syscall.Errno(0xc037010e))  // false before, true after

Add a regression test covering an errors.Is match on an unwrapped rpcError.

+++ internal/gcs/bridge.go
@@ -235,7 +235,13 @@ func (err *rpcError) Error() string {
 }
 
 func (err *rpcError) Unwrap() error {
-	return windows.Errno(err.result)
+	// result is an int32 HRESULT. HCS error codes have the high bit set
+	// (e.g. 0xc037010e), so a direct conversion to Errno (uintptr) would
+	// sign-extend the negative int32 to 0xFFFFFFFF_C037010E and no longer
+	// match the canonical syscall.Errno constants used with errors.Is
+	// (e.g. hcs.ErrComputeSystemDoesNotExist). Convert via uint32 to keep
+	// the code zero-extended.
+	return windows.Errno(uint32(err.result))
 }
 
 // Err returns the RPC's result. This may be a transport error or an error from
